requirements.txt
flask>=3.1.1
gunicorn>=23.0.0
google-generativeai>=0.8.0
twilio>=9.6.3
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.10
email-validator>=2.2.0
werkzeug>=3.1.3

# LangChain for RAG functionality
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0